requirements_38.txt
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
contourpy==1.1.1
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
MAVProxy==1.8.71
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pathlib2==2.3.7.post1
pillow==10.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pymavlink==2.4.41
pynmeagps==1.0.42
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.10.1
six==1.16.0
stringzilla==3.10.5
sympy==1.13.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.9.0
wxPython==4.2.2
